Factor out a main.js
