+++ src/index.html
@@ -10,7 +10,7 @@
   <body>
     <div id="schedule"></div>
     <div id="reload"></div>
-    <script src="./lib/remnote-schedule.js"></script>
+    <script src="./main.js"></script>
     <script>
       // TODO: Enable this only in deployment, not in local development.
       // if (window.self == window.top) {
